index.html

<!DOCTYPE html>
<html lang="en"> 
<head> 
  <title>W3C Testing How To</title> 
  <meta charset="utf-8">
  <meta name="copyright" 
   content="Copyright &#169; 2012  W3C. This work is distributed under the W3C® Software License." >
  <meta name="rights" content="see http://www.w3.org/Consortium/Legal/2002/copyright-software-20021231">
  <link rel="stylesheet" type="text/css" media="screen, projection, print" 
   href="http://www.w3.org/Talks/Tools/Slidy2/styles/slidy.css"> 
  <script src="http://www.w3.org/Talks/Tools/Slidy2/scripts/slidy.js"></script> 
  <style> 
   @media screen and (min-width: 700px) {
     div.slide {
       margin: 5%;
     }
   }
  </style> 
</head>
<body>
<div class="slide cover">
 <h1 style='margin-top: 15%'>W3C Testing How To</h1> 

 <p><a href='https://github.com/w3c/testing-how-to'>https://github.com/w3c/testing-how-to</a></p>
</div> 

<div class="slide">
 <h1>Content</h1>
 <ol>
   <li>What do you test in a specification?</li>
   <li>How to write a test?</li>
   <li>W3C Test repositories</li>
   <li>How to add a test suite into the W3C framework?</li>
   <li>Licenses For W3C Test Suites</li>
   <li>Strategies for testing a specification</li>
 </ol>
</div>

<div class="slide">
 <h1>What do you test in a specification?</h1>

 <ol>
  <li>Look for <i>conformance requirements</i> in the spec
  <li>Some statements are unambiguously requirements
  <li>Some statements are <i>candidate</i> requirements
  <li>Requirements are often stated within algorithms
  <li>A single requirement usually requires multiple test
  <li>A single test case may test a combination of requirements
  <li>In short, it’s complicated :-)
 </ol>
</div>

<div class="slide">
 <h1>Look for <i>conformance requirements</i> in the spec</h1>
 <p>
  In the simplest case, a <b>conformance requirement</b> is any single statement
  that uses RFC 2119 language (<i>MUST</i>, <i>MUST NOT</i>, <i>SHOULD</i>,
  <i>SHOULD NOT</i>, <i>MAY</i>), that specifies required behavior for a
  particular conformance class defined elsewhere in the spec (e.g., a Web
  browser or other user agent). However, some conformance requirements may not
  use RFC 2119 language.
 <p>
  Here’s a paragraph from the HTML5 spec that contains several
  conformance requirements:
 <blockquote>
  <p>The
   <code>dropEffect</code>
  attribute controls the drag-and-drop feedback that the user is given
  during a drag-and-drop operation.
  When the <code>DataTransfer</code> object is created, the
  <code>dropEffect</code> attribute is set to a string value.
  On getting, it must return its current value.
  On setting, if the new value is one of
  “<code title="">none</code>”,
  “<code title="">copy</code>”,
  “<code title="">link</code>”, or
  “<code title="">move</code>”, then the attribute’s current value
  must be set to the new value.
  Other values must be ignored.</p>
 </blockquote>
</div>
<div class="slide">
 <h1>Some statements are unambiguously requirements</h1>
 <p>
  Here again is a paragraph from the HTML5 spec that contains
  several conformance requirements:
 <blockquote>
  <p>The
   <code>dropEffect</code>
  attribute controls the drag-and-drop feedback that the user is given
  during a drag-and-drop operation.
  When the <code>DataTransfer</code> object is created, the
  <code>dropEffect</code> attribute is set to a string value.
  On getting, it must return its current value.
  On setting, if the new value is one of
  “<code title="">none</code>”,
  “<code title="">copy</code>”,
  “<code title="">link</code>”, or
  “<code title="">move</code>”, then the attribute’s current value
  must be set to the new value.
  Other values must be ignored.</p>
 </blockquote>
 <p>
  That paragraph contains three statements that use RFC 2119 “must”
  language and so are unambiguously requirements that need testing:
  <ol>
   <li><i>On getting, it must return its current value.</i>
   <li><i>On setting, if the new value is one of
    “<code title="">none</code>”,
    “<code title="">copy</code>”,
    “<code title="">link</code>”, or
    “<code title="">move</code>”, then the attribute’s current value
    must be set to the new value.</i>
   <li><i>Other values must be ignored.</i>
  </ol>
</div>
<div class="slide">
 <h1>Some statements are <i>candidate</i> requirements</h1>
 <p>
  Here again is a paragraph from the HTML5 spec that contains
  several conformance requirements:
 <blockquote>
  <p>The
   <code>dropEffect</code>
  attribute controls the drag-and-drop feedback that the user is given
  during a drag-and-drop operation.
  When the <code>DataTransfer</code> object is created, the
  <code>dropEffect</code> attribute is set to a string value.
  On getting, it must return its current value.
  On setting, if the new value is one of
  “<code title="">none</code>”,
  “<code title="">copy</code>”,
  “<code title="">link</code>”, or
  “<code title="">move</code>”, then the attribute’s current value
  must be set to the new value.
  Other values must be ignored.</p>
 </blockquote>
 <p>
  That paragraph contains one statement that may or may not actually be a
  requirement that needs testing:
 <blockquote>
  <p><i>When the <code>DataTransfer</code> object is
  created, the <code>dropEffect</code> attribute is set to a string value.</i>
 </blockquote>
  <p>Although that statement does not use RFC 2119 language, it still
  states something that could possibly be tested. Determining whether or
  not it <i>should</i> be tested is something that requires reading other
  parts of the spec. For this particular example, no test is necessary for
  the statement above, because the actual requirement for setting the value
  of the <code>dropEffect</code> when the <code>DataTransfer</code> object
  is created is defined in another part of the spec.
 </blockquote>
</div>
<div class="slide">
 <h1>Requirements are often stated within algorithms</h1>
 <p>
  Some algorithms in a spec may contain many individual statements or steps
  that don’t use each use RFC 2119 but that nonetheless state conformance
  requirements. Such algorithms are typically prefaced by a statement that
  does use RFC 2119 language.
 <p>Here’s an excerpt from an algorithm in the HTML5 spec:
 <blockquote>
  <p>When a user agent is required to <dfn id="fire-a-dnd-event">fire a DND event</dfn>
    named <var title="">e</var> at an element, using a particular
    drag data store, the user agent must run the following
    steps:</p>
   <ol>

   <li>
   <p>Let <var title="">dataTransfer</var> be a newly created
   <code>DataTransfer</code> object associated with the given
   drag data store.

   <li>
   <p>Set the <code>effectAllowed</code>
   attribute to the drag data store's drag data store allowed effects state.

   <li>
    <p>Set the <code>dropEffect</code> attribute to
    "<code>none</code>" if <var>e</var> is <code>dragstart</code>, <code>drag</code>, or <code>dragleave</code>…
  </ol>
  </blockquote>
  <p>Notice that the algorithm is prefaced by a statement saying, “the user
   agent must run the following steps”.
</div>

<div class="slide">
 <h1>A single requirement usually needs multiple tests</h1>
 <p>
  There usually isn’t a one-to-one relationship between a conformance
  requirement in a spec and the number of test cases needed to test
  that requirement adequately. Instead, each requirement typically needs
  multiple test cases.
  For example, the following statement from the HTML5 spec would need at
  least four test cases—one for each of the possible values given.
 <blockquote>
  <p>
  On setting, if the new value is one of
  “<code title="">none</code>”,
  “<code title="">copy</code>”,
  “<code title="">link</code>”, or
  “<code title="">move</code>”,
  then the attribute’s current value must be set to the new value.
 </blockquote>
 <p>
  That requirement likely in fact would actually need even more test cases.
  Determining how to test it adequately would require reading other parts
  of the spec.
</div>

<div class="slide">
 <h1>A test case may test a combination of statements</h1>
 <p>
  It’s always preferable for a test case to focus as narrowly as
  possible on isolating a test for one specific case (though that case will
  almost always have dependencies on other major features: the parsing
  algorithm, the entire DOM, etc.) But often it’s not that simple to narrow
  it down that specifically. For example, in the case of testing complex
  algorithms, a single test case may necessarily be testing a combination of
  requirements from different parts of the algorithm.
</div>

<div class="slide">
 <h1>In short, it’s complicated :-)</h1>
 <p>
  For any particular spec, determining things like which statements in the
  spec require testing and how many test cases are needed to test them is
  something that requires a very close reading of the spec and very close
  familiarity with it.
 <p>
  Different spec editors take different approaches—and though the style and
  conventions of W3C specs for browser technologies have “normalized” and
  improved considerably over the years in terms of the degree of rigor and
  precision and lack of ambiguity with which the specs state requirements
  (for example, using RFC 2119 language consistently to clearly identify
  requirements), isolating exactly what to test and how to test it
  necessitates first spending a lot of time reading the spec.
</div>

<div class="slide">
 <h1>How to write a test?</h1>

 <ol type="lower-alpha">
  <li>Script Test</li>
  <li>Reftests</li>
  <li>Self-Describing Test</li>
  <li>Client-Server Test</li>
 </ol>
</div>

<div class='slide'>
<h1>Script Test</h1>
<p>When the result of your test could be determined by a script.</p>
<pre style='font-size: larger'>&lt;script
  src="<a href='http://w3c-test.org/resources/testharness.js'>/resources/testharness.js</a>">
&lt;/script>
&lt;script
  src="<a href='http://w3c-test.org/resources/testharnessreport.js'>/resources/testharnessreport.js</a>">
&lt;/script></pre>

</div>

<div class='slide'>
<h1>testharness.js</h1>

<pre>
&lt;p id="MyE">This test checks the
  &lt;span>firstElementChild&lt;/span> property.&lt;/p>
&lt;div <em style='color: green'>id='log'</em>>&lt;/div>
&lt;script>
<em style='color: green'>test</em>(
 function() {
  var myE = document.getElementById("MyE");
  <em style='color: green'>assert_true</em>(!!MyE.firstElementChild);
 },
 "Element.firstElementChild exists"
);
&lt;/script>
</pre>

<p>See <a href='http://w3c-test.org/webperf/tests/approved/navigation-timing/html5/test_timing_attributes_exist.html'>Demo</a></p>
</div>

<div class='slide'>
<h1>testharness.js: assertions</h1>
<p>
<code>assert_true</code>,
<code>assert_false</code>,
<code>assert_equals</code>,
<code>assert_not_equals</code>,
<code>assert_approx_equals</code>
</p>
<p>
<code>assert_in_array</code>,
<code>assert_array_equals</code>,
<code>assert_regexp_equals</code>
</p>
<p>
<code>assert_own_property</code>,
<code>assert_inherits</code>,
<code>assert_idl_attribute</code>,
<code>assert_readonly</code>,
<code>assert_throws</code>,
<code>assert_unreached</code>,
<code>assert_any</code>
</p>
</div>

<div class='slide'>
<h1>testharness.js: async test</h1>

<pre>
var t = <em style='color: green'>async_test</em>("Element.firstElementChild exists");

function run() {
  t.<em style='color: green'>step</em>(
    function () {
      assert_true(!!MyE.firstElementChild);
    }
  );
  t.<em style='color: green'>done</em>();
};
window.onload = run;
</pre>

<p>See <a href='http://w3c-test.org/webapps/WebMessaging/tests/submissions/Microsoft/event.data.htm'>Demo</a> @@find a better example...</p> 
</div>

<div class='slide'>
<h1>testharness.js: metadata</h1>

<p>In the &lt;head> element:
<pre>
&lt;<em style='color: green'>title</em>>Element.firstElementChild exists&lt;/title&gt;
&lt;link rel="<em style='color: green'>help</em>"
 href="http://www.w3.org/TR/domcore/#dom-element-firstelementchild"/>
&lt;meta name="<em style='color: green'>assert</em>"
 content="The firstElementChild property must exist"/>
&lt;link rel="<em style='color: green'>author</em>"
 title="John Doe" href="mailto:john.doe@example.com"/>
</pre>
</div>

<div class='slide'>
<h1>testharness.js: per-test metadata</h1>

<pre>&lt;script>
test(
  function() {
    var myE = document.getElementById("MyE");
    assert_true(!!MyE.firstElementChild);
  },
  "Element.firstElementChild exists",
  {
    <em style='color: green'>help</em>: 'http://www.w3.org/TR/domcore/#dom-element-firstelementchild', 
    <em style='color: green'>assert</em>: ['The firstElementChild property must exist'],
    <em style='color: green'>author</em>: 'John Doe &lt;john.doe@example.com>'
  }
);
&lt;/script></pre>
</div>

<div class='slide'>
<h1>testharness.js: further readings</h1>

<ul>
<li>Read the <a href='http://w3c-test.org/resources/testharness.js'>documentation in the source</a></li>
<li><a href='http://darobin.github.com/test-harness-tutorial/docs/using-testharness.html'>testharness.js tutorial</a></li>
<li><a href='http://wiki.csswg.org/test/scripttest'>Writing Script Tests</a></li>
</ul>
</div>

<div class='slide'>
<h1>Reftest</h1>

<p>When the result of your test is visual.</p>

<table id='cssFirstLetter'>
<tr>
<th>bdo-000.html</th>
<th>bdo-000-ref.html</th>
</tr>
<tr class='lines'>
<td style='width: 45%'>
<bdo dir="rtl">HEBREW</bdo>
</td>
<td style='width: 45%'>
<span dir="ltr">WERBEH</span>
</td>
</tr>
<tr>
<td>
<pre>
&lt;bdo dir="rtl">HEBREW&lt;/bdo>
</pre>
</td>
<td>
<pre>
&lt;span>WERBEH&lt;/span>
</pre>
</td>
</tr>
</table>
</div>

<div class='slide'>
<h1>reftest: metadata</h1>

<p>As for testharness.js, put it in the &lt;head> element:
<pre>
&lt;<em style='color: green'>title</em>>rtl value on dir attribute on bdo element&lt;/title&gt;
&lt;link rel="<em style='color: green'>help</em>"
 href="http://www.w3.org/TR/html5/global-attributes.html#attr-dir-rtl"/>
&lt;link rel="<em style='color: green'>help</em>"
 href="http://www.w3.org/TR/html5/the-bdo-element.html#the-bdo-element"/>
&lt;meta name="<em style='color: green'>assert</em>"
 content="The dir attribute on bdo element could invert the text direction"/>
&lt;link rel="<em style='color: green'>author</em>"
 title="John Doe" href="mailto:john.doe@example.com"/>
</pre>
</div>

<div class='slide'>
<h1>Reftest: further readings</h1>

  <ul>
   <li><a href='http://wiki.csswg.org/test/reftest'>Writing Reftests</a></li>
   <li><a href='http://wiki.csswg.org/test/format#template-for-new-tests'>Metadata for Reftests</a></li>
  </ul>
</div>

<div class='slide'>
<h1>Manual test</h1>

<p>When everything else wouldn't work.</p>

<div>
 <h2>visibilitychange</h2>
 <p>Iconify this window and restore it.</p>
 <div id='visibility' style='width: 50px; height: 50px;background-color: red;margin: auto;'></div>

 <p>If the square is still red, please indicate that the test has failed.</p>

<script>
document.addEventListener("visibilitychange",
 function() {
   document.getElementById("visibility").style.backgroundColor = 'green';
 }, false);
</script>

<p>See also <a href='http://w3c-test.org/webperf/tests/submission/Microsoft/PageVisibility/test_minimize.htm'>Demo</a></p>

</div>

</div>

<div class="slide">
 <h1>Client-Server test</h1>

 <p>When you need server-side code as well</p>

 <p>So far, we have two examples of those:
 <ol>
  <li>XMLHttpRequest</li>
  <li>Web Sockets</li>
 </ol>
</div>

<div class="slide">
 <h1>XHR test</h1>

 <p>Use <a href='https://dvcs.w3.org/hg/webapps/file/tip/XMLHttpRequest/tests/submissions/Opera/resources'>PHP</a> to simulate server responses:</p>

 <ol>
   <li><a href="https://dvcs.w3.org/hg/webapps/file/tip/XMLHttpRequest/tests/submissions/Opera/resources/status.php">status.php</a> for content, http code, and media type (see <a href='http://w3c-test.org/webapps/XMLHttpRequest/tests/submissions/Opera/resources/status.php?content=%3Cx%3E%3C/x%3E&type=application/xml'>example</a>)</li>
   <li><a href="https://dvcs.w3.org/hg/webapps/file/tip/XMLHttpRequest/tests/submissions/Opera/resources/delay.php">delay.php</a> to delay HTTP responses</li>
   <li>&hellip;</li>
</ol>
</div>

<div class="slide">
 <h1>Web Sockets</h1>

 <p>We currently have two Web Socket servers running on w3c-test.org:

 <ul>
  <li>A mod_pywebsocket-based server integrated into the main Apache Web
    server, and so listening on ports 80 and 443. (There are not yet any
    test cases that use that instance, but there is an example at
    <a href="http://w3c-test.org/demo/console.html">http://w3c-test.org/demo/console.html</a>;
    writing test cases that use that instance requires installing
    additional *_wsh.py handlers on the server.)
  <li>A simple “echo” server (Jetty-based) listening on ports 8080 and 8081
    and 8443 (secure); to use this server, your test case needs to include the
    <a href="http://w3c-test.org/webapps/WebSockets/tests/submissions/Microsoft/websocket.js">http://w3c-test.org/webapps/WebSockets/tests/submissions/Microsoft/websocket.js</a>
    script; for further information, view the source of that script; some
    existing test cases are at
    <a href="http://w3c-test.org/webapps/WebSockets/tests/submissions/Microsoft/">http://w3c-test.org/webapps/WebSockets/tests/submissions/Microsoft/</a>
  </ul>
</div>

<div class='slide'>
  <h1>Audio and video materials</h1>

  <p>If you're testing audio or video content, we have content already:</p>

  <ol>
    <li><a href='http://media.w3.org/2010/05/'>2010/05</a>: Open
    project movies, time ticker sounds, timed videos, single frame
      videos, etc.</li>
    <li><a href='http://media.w3.org/2011/01/'>2011/01</a>: sounds, timed videos</li>
    <li><a href='http://media.w3.org/2011/09/audio'>2011/09</a>: multi audio tracks</li>
  </ol>
</div>


<div class="slide">
 <h1>W3C test-case repositories</h1>

 <p>We currently have two shared areas for repositories for storing test cases:

 <ol>
  <li>A large set of mercurial repos at <a href="http://dvcs.w3.org/">http://dvcs.w3.org/</a>
  <li>A smaller (currently) set of git repos at <a href="https://github.com/w3c/">https://github.com/w3c/</a>
 </ol>

 <p>If you don’t have admin access to create new repos in either of those
  places, contact <a href="mailto:mike@w3.org">Mike Smith &lt;mike@w3.org></a>
  with the details and he’ll set up the repo for you.

 <p>The repos are all ACL-controlled, so submit (push) new test cases into
  a particular repo, you either need to added to the list of authorized
  contributors, or you need to be a member of a working group all of whose
  members are authorized to push to the repo.

 <p>If you need push access to a particular repo, contact
  <a href="mailto:mike@w3.org">Mike Smith &lt;mike@w3.org></a>.

 <p>The general protocol for adding new test cases is to place your tests
  in a subdirectory that has either your company’s name or your personal
  name, under the top-level <code>submission</code> directory in the repo.
</div>


<div class="slide">
 <h1>Maintaining a Test suite</h1>

 <ul>
   <li>How to add a test suite into the W3C framework</li>
   <li>@@TODO modify a test suite?</li>
   <li>@@TODO remove a test suite?</li>
   <li>@@TODO remove test suite results?</li>
 </ul>

</div>

<div class="slide">
 <h1>How to add a test suite into the W3C framework</h1>

 <p>We have a shared system at
   <a href="http://w3c-test.org/framework/">http://w3c-test.org/framework/</a>
   for running test suites
 <p>Documentation on adding a new test suite to the framework is available at
  <a href="http://w3c-test.org/framework/docs/maintainer/">http://w3c-test.org/framework/docs/maintainer/</a>.
 <p>There are four basic steps to adding a new test suite:

  <ol>
   <li>Add metadata for the specification the test suite is associated with.
   <li>Add metadata about the sections/structure of that specification.
   <li>Add metadata about the overall test suite.
   <li>Add metadata about all the test cases in that test suite.
  </ol>

 <p>Once you’ve completed those steps, anyone will be able to come in and
  run the actual test cases in their own browsers.
</div>

<div class="slide">
 <h1>Test suite results</h1>

 <ul>
   <li>Test suite results may be submitted automatically or manually to the W3C framework</li>
   <li>Results are stored and be accessed on w3c-test.org</li>
   <li>See also the <a href='http://w3c-test.org/framework/docs/api/'>Test Framework JSON API</a></li>
 </ul>
</div>

<div class="slide">
 <h1>Licenses For W3C Test Suites</h1>

  <p>W3C makes test suites available under two distinct licenses:
<ol>
<li>a <a href="http://www.w3.org/Consortium/Legal/2008/04-testsuite-license.html">W3C Test Suite License</a> for
    an Authoritative W3C Test Suite or when claims of performance with
    respect to a specification are required.</li>
<li>a <a href="http://www.w3.org/Consortium/Legal/2008/03-bsd-license.html">3-clause
BSD License</a> for other applications.</li>
</ol>

</div>

<div class="slide">
 <h1>Contributions For W3C Test Suites</h1>

 <ol>
   <li>If you're a participant of the Working Group, you already
   agreed to the requirements, so no special action is required in
   order to contribute.</li>
   <li>otherwise, as a non-participant, use one of the following agreements:
     <ul>
       <li><a href="http://www.w3.org/2002/09/wbs/1/testgrants-200409/">Grant&#xA0;I</a> for contributed test cases published <strong>as part of</strong> a W3C Recommendation</li>
       <li><a href="http://www.w3.org/2002/09/wbs/1/testgrants2-200409/">Grant&#xA0;II</a> for contributed test cases published <strong>outside</strong> a W3C Recommendation</li>
     </ul>
   </li>
 </ol>

</div>

<div class="slide">
 <h1>Strategies for testing a specification</h1>

 <p>@@Guidelines for spec authors?</p>

</div>

<div class="slide">
 <h1>For more information</h1>

 <p>Use <a href='http://lists.w3.org/Archives/Public/public-test-infra/'>public-test-infra@w3.org</a></p>

</div>

</body>
</html>
